% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_de.R
\name{acceptable-de}
\alias{acceptable-de}
\alias{.acceptablePValueFields}
\alias{.acceptableLogFCFields}
\alias{.acceptableAveAbFields}
\alias{.setAcceptableFields}
\title{Acceptable fields for DE panels}
\usage{
.acceptablePValueFields()

.acceptableLogFCFields()

.acceptableAveAbFields()

.setAcceptableFields(PValue = NULL, LogFC = NULL, AveAb = NULL)
}
\arguments{
\item{PValue, LogFC, AveAb}{Character vector of acceptable fields (usually in the \code{\link{rowData}}) for a given statistic.}
}
\value{
\code{.acceptablePValueFields} will return a character vector of acceptable names for p-value fields.

\code{.acceptableLogFCFields} will return a character vector of acceptable names for log-FC fields.

\code{.acceptableAveAbFields} will return a character vector of acceptable names for average abundance fields.

\code{.setAcceptableFields} will define the set of acceptable fields.
}
\description{
Set or get the acceptable fields to use for all \linkS4class{Panel} instances related to differential expression,
including \linkS4class{VolcanoPlot} and \linkS4class{MAPlot}.
}
\examples{
old <- .acceptablePValueFields()
old

.setAcceptableFields(PValue="YAY")
.acceptablePValueFields()

# Restoring.
.setAcceptableFields(PValue=old)

}
\author{
Aaron Lun
}
