% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{.getAcceptablePValueFields}
\alias{.getAcceptableAveAbFields}
\alias{.getAcceptableLogFCFields}
\alias{.setAcceptablePValueFields}
\alias{.setAcceptableAveAbFields}
\alias{.setAcceptableLogFCFields}
\title{Defunct functions}
\usage{
.getAcceptablePValueFields(...)

.getAcceptableAveAbFields(...)

.getAcceptableLogFCFields(...)

.setAcceptablePValueFields(...)

.setAcceptableAveAbFields(...)

.setAcceptableLogFCFields(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
All functions error out with a defunct message pointing towards its replacement (if available).
}
\description{
Pretty much as it says here. These functions are all defunct and should not be used.
}
\author{
Aaron Lun
}
