% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialStatisticsTable.R
\name{DifferentialStatisticsTable-class}
\alias{DifferentialStatisticsTable-class}
\alias{DifferentialStatisticsTable}
\alias{initialize,DifferentialStatisticsTable-method}
\alias{.fullName,DifferentialStatisticsTable-method}
\alias{.panelColor,DifferentialStatisticsTable-method}
\alias{.defineDataInterface,DifferentialStatisticsTable-method}
\alias{.generateTable,DifferentialStatisticsTable-method}
\alias{.createObservers,DifferentialStatisticsTable-method}
\alias{.multiSelectionInvalidated,DifferentialStatisticsTable-method}
\title{Differential statistics table}
\description{
A table that dynamically computes differential statistics based on a selected subset of samples.
Comparisons are made between the active selection in the transmitting panel
and (i) all non-selected points, if no saved selections are available;
or (ii) each subset of points in each saved selection.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualization:
\itemize{
\item \code{LogFC}, a numeric scalar indicating the log-fold change threshold to test against.
\item \code{TestMethod}, string indicating the test to use (based on the \code{findMarkers} function from \pkg{scran}).
This can be \code{"t"}, \code{"wilcox"} or \code{"binom"}.
}

In addition, this class inherits all slots from its parent \linkS4class{RowTable},
\linkS4class{Table} and \linkS4class{Panel} classes.
}

\section{Constructor}{

\code{DifferentialStatisticsTable(...)} creates an instance of a DifferentialStatisticsTable class,
where any slot and its value can be passed to \code{...} as a named argument.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{DifferentialStatisticsTable} class.
Refer to the documentation for each method for more details on the remaining arguments.

For defining the interface:
\itemize{
\item \code{\link{.defineDataInterface}(x, se, select_info)} returns a list of interface elements for manipulating all slots described above.
\item \code{\link{.panelColor}(x)} will return the specified default color for this panel class.
\item \code{\link{.fullName}(x)} will return \code{"Differential statistics table"}.
\item \code{\link{.hideInterface}(x)} will return \code{TRUE} for UI elements related to multiple row selections,
otherwise calling the method for \linkS4class{RowTable}.
}

For monitoring reactive expressions:
\itemize{
\item \code{\link{.createObservers}(x, se, input, session, pObjects, rObjects)} sets up observers for all new slots described above, as well as in the parent classes via the \linkS4class{RowTable} method.
}

For creating the table:
\itemize{
\item \code{\link{.generateTable}(x, envir)} will create a data.frame of newly computed statistics in \code{envir}.
The method will return the commands required to do so as well as a list of labels.
}
}

\examples{
library(scRNAseq)
library(scater)

sce <- ReprocessedAllenData(assays="tophat_counts")
sce <- logNormCounts(sce, exprs_values="tophat_counts")
sce <- runPCA(sce, ncomponents=4)
sce <- runTSNE(sce)

dst <- DifferentialStatisticsTable(PanelId=1L, PanelWidth=8L,
    ColumnSelectionSource="ReducedDimensionPlot1")

rdp <- ReducedDimensionPlot(PanelId=1L, 
    ColorByFeatureSource="DifferentialStatisticsTable1")

if (interactive()) {
    iSEE(sce, initial=list(rdp, dst))
}

}
