% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{getFeatureSetCommands}
\alias{getFeatureSetCommands}
\alias{setFeatureSetCommands}
\title{Global feature set commands}
\usage{
getFeatureSetCommands()

setFeatureSetCommands(value)
}
\arguments{
\item{value}{A list of two character vectors named \code{"CreateCollections"} and \code{"RetrieveSet"}.
Both vectors should be of the same length and have the same names.
Vectors should contain R commands to create collections and retrieve sets;
see \code{?\linkS4class{FeatureSetTable}} and the output of \code{\link{createGeneSetCommands}} for details.}
}
\value{
\code{getFeatureSetCommands} returns the current global feature set commands.

\code{setFeatureSetCommands} will set the current global feature set commands and return \code{NULL} invisibly.
}
\description{
Get or set the commands to define the global collection of feature sets.
}
\details{
These utilities allow users to easily set the feature set commands for all \linkS4class{FeatureSetTable}s at once.
Note that they work during construction of each class instance and has no effect on the \code{iSEE} application once it starts.
}
\examples{
old <- getFeatureSetCommands()

new.cmds <- createGeneSetCommands(organism="org.Mm.eg.db", 
    identifier="SYMBOL")
setFeatureSetCommands(new.cmds)

getFeatureSetCommands()

setFeatureSetCommands(old) 
}
\seealso{
\code{\link{createGeneSetCommands}}, for one method of generating \code{value}.
}
\author{
Aaron Lun
}
