% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{getPValueFields}
\alias{getPValueFields}
\alias{getLogFCFields}
\alias{getAveAbFields}
\alias{setPValueFields}
\alias{setLogFCFields}
\alias{setAveAbFields}
\title{Global DE fields}
\usage{
getPValueFields()

getLogFCFields()

getAveAbFields()

setPValueFields(value)

setLogFCFields(value)

setAveAbFields(value)
}
\arguments{
\item{value}{A character vector containing the names of acceptable columns for each statistic.}
}
\value{
\code{getPValueFields} returns the acceptable column names for p-values.

\code{getLogFCFields} returns the acceptable column names for log-fold changes.

\code{getAveAbFields} returns the acceptable column names for the average abundances.

The corresponding setters set the global column names for each statistic and return \code{NULL} invisibly.
}
\description{
Get or set the names of \code{\link{rowData}} columns related to a differential expression analysis.
}
\examples{
old <- getPValueFields()

setPValueFields(LETTERS)
getPValueFields()

setPValueFields(old)
}
\seealso{
\linkS4class{VolcanoPlot} and \linkS4class{MAPlot}, which are affected by these globals.
}
\author{
Aaron Lun
}
