% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEu-package.R
\docType{package}
\name{iSEEu-pkg}
\alias{iSEEu}
\alias{iSEEu-package}
\title{iSEEu: iSEE Universe}
\description{
\code{iSEEu} is a package that provides modes and panels for \code{iSEE}, allowing easy configuration of \code{iSEE} applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iSEE/iSEEu}
  \item Report bugs at \url{https://github.com/iSEE/iSEEu/issues}
}

}
\author{
Charlotte Soneson \email{charlottesoneson@gmail.com}

Federico Marini \email{marinif@uni-mainz.de}

Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk}

Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}
}
\keyword{internal}
