% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot.R
\docType{methods}
\name{MAPlot-class}
\alias{MAPlot-class}
\alias{MAPlot}
\alias{initialize,MAPlot-method}
\alias{.cacheCommonInfo,MAPlot-method}
\alias{.refineParameters,MAPlot-method}
\alias{.defineDataInterface,MAPlot-method}
\alias{.createObservers,MAPlot-method}
\alias{.hideInterface,MAPlot-method}
\alias{.fullName,MAPlot-method}
\alias{.panelColor,MAPlot-method}
\alias{.generateDotPlotData,MAPlot-method}
\alias{.allowableXAxisChoices,MAPlot-method}
\alias{.allowableYAxisChoices,MAPlot-method}
\alias{.prioritizeDotPlotData,MAPlot-method}
\alias{.colorByNoneDotPlotField,MAPlot-method}
\alias{.colorByNoneDotPlotScale,MAPlot-method}
\alias{.generateDotPlot,MAPlot-method}
\alias{.definePanelTour,MAPlot-method}
\title{The MAPlot class}
\description{
The MAPlot is a \linkS4class{RowDataPlot} subclass that is dedicated to creating a MA plot.
It retrieves the log-fold change and average abundance and creates a row-based plot where each point represents a feature.
Users are expected to load relevant statistics into the \code{\link{rowData}} of a \linkS4class{SummarizedExperiment}.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualization:
\itemize{
\item \code{PValueField}, a string specifying the field of \code{\link{rowData}} containing the p-values.
\item \code{PValueThreshold}, a numeric scalar in (0, 1] specifying the threshold to use on the (adjusted) p-value.
Defaults to 0.05.
\item \code{LogFCThreshold}, a non-negative numeric scalar specifying the threshold to use on the log-fold change.
Defaults to 0.
\item \code{PValueCorrection}, a string specifying the multiple testing correction to apply.
Defaults to \code{"BH"}, but can take any value from \code{\link{p.adjust.methods}}.
}

The following slots control the choice of columns in the user interface:
\itemize{
\item \code{PValuePattern}, a character vector specifying the patterns of all potential columns containing p-values, see \code{\link{getPValuePattern}}.
\item \code{LogFCPattern}, a character vector specifying the patterns of all potential columns containing log-fold changes, see \code{\link{getLogFCPattern}}.
\item \code{AveAbPattern}, a character vector specifying the patterns of all potential columns containing average abundances, see \code{\link{getAveAbPattern}}.
}

In addition, this class inherits all slots from its parent \linkS4class{RowDataPlot},
\linkS4class{RowDotPlot}, \linkS4class{DotPlot} and \linkS4class{Panel} classes.
}

\section{Constructor}{

\code{MAPlot(...)} creates an instance of a MAPlot class,
where any slot and its value can be passed to \code{...} as a named argument.

Initial values for \code{PValuePattern}, \code{AveAbPattern} and \code{LogFCPattern} are set to the outputs of \code{\link{getPValuePattern}}, \code{\link{getAveAbPattern}} and \code{\link{getLogFCPattern}}, respectively.
These parameters are considered to be global constants and cannot be changed inside the running \code{iSEE} application.
Similarly, it is not possible for multiple MAPlots in the same application to have different values for these slots;
within the app, all values are set to those of the first encountered MAPlot to ensure consistency.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{RowDataPlot} class.
Refer to the documentation for each method for more details on the remaining arguments.

For setting up data values:
\itemize{
\item \code{\link{.cacheCommonInfo}(x, se)} returns \code{se} after being loaded with class-specific constants.
This includes \code{"valid.p.fields"}, \code{"valid.ab.fields"} and \code{"valid.lfc.fields"}, which are character vectors containing the names of valid \code{\link{rowData}} columns for the p-values, average abundances and log-fold changes, respectively.
\item \code{\link{.refineParameters}(x, se)} returns \code{x} after setting \code{XAxis="Row data"} and the various \code{*Pattern} fields to their cached values.
This will also call the equivalent \linkS4class{RowDataPlot} method for further refinements to \code{x}.
If valid p-value, abundance and log-fold change fields are not available, \code{NULL} is returned instead.
}

For defining the interface:
\itemize{
\item \code{\link{.defineDataInterface}(x, se, select_info)} returns a list of interface elements for manipulating all slots described above.
\item \code{\link{.panelColor}(x)} will return the specified default color for this panel class.
\item \code{\link{.allowableXAxisChoices}(x, se)} returns a character vector specifying the acceptable average abundance-related variables in \code{\link{rowData}(se)} that can be used as choices for the x-axis.
\item \code{\link{.allowableYAxisChoices}(x, se)} returns a character vector specifying the acceptable log-fold change-related variables in \code{\link{rowData}(se)} that can be used as choices for the y-axis.
\item \code{\link{.hideInterface}(x, field)} will return \code{TRUE} for \code{field="XAxis"},
otherwise it will call the \linkS4class{RowDataPlot} method.
\item \code{\link{.fullName}(x)} will return \code{"MA plot"}.
}

For monitoring reactive expressions:
\itemize{
\item \code{\link{.createObservers}(x, se, input, session, pObjects, rObjects)} sets up observers for all new slots described above, as well as in the parent classes via the \linkS4class{RowDataPlot} method.
}

For creating the plot:
\itemize{
\item \code{\link{.generateDotPlotData}(x, envir)} will create a data.frame of row metadata variables in \code{envir}.
This should contain average abundances on the x-axis and log-fold changes on the y-axis,
in addition to an extra field specifying whether or not the feature was considered to be significantly up or down.
The method will return the commands required to do so as well as a list of labels.
\item \code{\link{.prioritizeDotPlotData}(x, envir)} will create variables in \code{envir} marking the priority of points.
Significant features receive higher priority (i.e., are plotted over their non-significant counterparts) and are less aggressively downsampled when \code{Downsample=TRUE}.
The method will return the commands required to do this as well as a logical scalar indicating that rescaling of downsampling resolution is performed.
\item \code{\link{.colorByNoneDotPlotField}(x)} will return a string specifying the field of the data.frame (generated by \code{\link{.generateDotPlotData}}) containing the significance information.
This is to be used for coloring when \code{ColorBy="None"}.
\item \code{\link{.colorByNoneDotPlotScale}(x)} will return a string containing a \pkg{ggplot2} command to add a default color scale when \code{ColorBy="None"}.
\item \code{\link{.generateDotPlot}(x, labels, envir)} returns a list containing \code{plot} and \code{commands}, using the inital \linkS4class{ColumnDataPlot} \link{ggplot} and adding horizontal lines demarcating the log-fold change threshold.
}

For documentation:
\itemize{
\item \code{\link{.definePanelTour}(x)} returns an data.frame containing the steps of a panel-specific tour.
}
}

\examples{
# Making up some results:
se <- SummarizedExperiment(matrix(rnorm(10000), 1000, 10))
rownames(se) <- paste0("GENE_", seq_len(nrow(se)))
rowData(se)$PValue <- runif(nrow(se))
rowData(se)$LogFC <- rnorm(nrow(se))
rowData(se)$AveExpr <- rnorm(nrow(se))

if (interactive()) {
    iSEE(se, initial=list(MAPlot()))
}

}
\seealso{
\link{RowDataPlot}, for the base class.
}
\author{
Aaron Lun
}
