% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeReducedDim.R
\name{modeReducedDim}
\alias{modeReducedDim}
\title{App pre-configured to compare multiple reduced dimension plots}
\usage{
modeReducedDim(
  se,
  includeNames = reducedDimNames(se),
  colorBy = NULL,
  ...,
  plot_width = 4
)
}
\arguments{
\item{se}{An object that coercible to \linkS4class{SingleCellExperiment}}

\item{includeNames}{Character vector with the names of reduced dimensions
to display as individual panels. The default uses all available in
\code{reducedDimNames(se)}.}

\item{colorBy}{Character scalar controling coloring of cells. Must match either
to one of \code{colnames(colData(se))} or \code{rownames(se)}.}

\item{...}{Additional arguments passed to \code{\link{iSEE}}.}

\item{plot_width}{The grid width of linked plots (numeric vector of length
either 1 or equal to \code{length(includeNames)}}
}
\value{
A Shiny app object is returned.
}
\description{
This mode launches a Shiny App preconfigured with multiple linked reduced
dimension plots for interactive data exploration of the
\code{\link{SingleCellExperiment}} object.
}
\examples{
library(scRNAseq)

# Example data ----
sce <- ReprocessedAllenData(assays="tophat_counts")
class(sce)

library(scater)
sce <- logNormCounts(sce, exprs_values="tophat_counts")
sce <- runPCA(sce, ncomponents = 30)
sce <- runTSNE(sce)
sce <- runUMAP(sce)
reducedDimNames(sce)

# launch the app ----
app <- modeReducedDim(sce, colorBy = "Primary.Type")
if (interactive()) {
  shiny::runApp(app, port=1234)
}
app <- modeReducedDim(sce, colorBy = "Scnn1a")
if (interactive()) {
  shiny::runApp(app, port=1234)
}
}
