% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_de.R
\name{utils-de}
\alias{utils-de}
\alias{.getAcceptablePValueFields}
\alias{.getAcceptableLogFCFields}
\alias{.getAcceptableAveAbFields}
\alias{.setAcceptablePValueFields}
\alias{.setAcceptableLogFCFields}
\alias{.setAcceptableAveAbFields}
\title{Acceptable fields for DE panels}
\usage{
.getAcceptablePValueFields()

.getAcceptableLogFCFields()

.getAcceptableAveAbFields()

.setAcceptablePValueFields(value)

.setAcceptableLogFCFields(value)

.setAcceptableAveAbFields(value)
}
\arguments{
\item{value}{Character vector of acceptable fields (usually in the \code{\link{rowData}}) for a given statistic.}
}
\value{
\code{.getAcceptablePValueFields} will return a character vector of acceptable names for p-value fields.

\code{.getAcceptableLogFCFields} will return a character vector of acceptable names for log-FC fields.

\code{.getAcceptableAveAbFields} will return a character vector of acceptable names for average abundance fields.

The setter functions will define the set of acceptable fields and return \code{NULL} invisibly.
}
\description{
Set or get the acceptable fields to use for all \linkS4class{Panel} instances related to differential expression,
including \linkS4class{VolcanoPlot} and \linkS4class{MAPlot}.
}
\examples{
old <- .getAcceptablePValueFields()
old

.setAcceptablePValueFields("YAY")
.getAcceptablePValueFields()

# Restoring.
.setAcceptablePValueFields(old)

}
\author{
Aaron Lun
}
