% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregatedDotPlot.R
\name{AggregatedDotPlot}
\alias{AggregatedDotPlot}
\alias{AggregatedDotPlot-class}
\alias{.cacheCommonInfo,AggregatedDotPlot-method}
\alias{.refineParameters,AggregatedDotplot-method}
\alias{.generateOutput,AggregatedDotPlot-method}
\alias{.renderOutput,AggregatedDotPlot-method}
\alias{.exportOutput,AggregatedDotPlot-method}
\alias{.defineDataInterface,AggregatedDotPlot-method}
\alias{.defineInterface,AggregatedDotPlot-method}
\alias{.hideInterface,AggregatedDotPlot-method}
\alias{.panelColor,AggregatedDotPlot-method}
\alias{.fullName,AggregatedDotPlot-method}
\title{The AggregatedDotPlot class}
\description{
Implements an aggregated dot plot where each feature/group combination is represented by a dot.
The color of the dot scales with the mean assay value across all samples for a given group,
while the size of the dot scales with the number of non-zero values across all samples.
}
\examples{
library(scRNAseq)

# Example data ----
sce <- ReprocessedAllenData(assays="tophat_counts")
class(sce)

library(scater)
sce <- logNormCounts(sce, exprs_values="tophat_counts")

# launch the app itself ----
if (interactive()) {
    iSEE(sce, initial=list(
        AggregatedDotPlot(ColumnData=c("Primary.Type"))
    ))
}

}
