% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{getTableExtraFields}
\alias{getTableExtraFields}
\alias{setTableExtraFields}
\title{Global extra table fields}
\usage{
getTableExtraFields()

setTableExtraFields(value)
}
\arguments{
\item{value}{A character vector containing the names of extra fields to include.}
}
\value{
\code{getTableExtraFields} returns the current global extra table fields.

\code{setTableExtraFields} will set the current global extra table fields and return \code{NULL} invisibly.
}
\description{
Get or set the names of the extra \code{\link{rowData}} or \code{\link{colData}} fields to include in a table.
}
\details{
By setting these values, all subsequent constructions of \linkS4class{DynamicMarkerTable} will have the same set of extra fields specified in their \code{"ExtraFields"} slot.
This allows users to easily customize all \linkS4class{DynamicMarkerTable} parameters at once.
Note that it only applies during the construction of the \linkS4class{DynamicMarkerTable} and has no effect on the \code{iSEE} application once it starts.
}
\examples{
old <- getTableExtraFields()

setTableExtraFields(LETTERS)
getTableExtraFields()

setTableExtraFields(old)
}
\author{
Aaron Lun
}
