% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{getTableExtraFields}
\alias{getTableExtraFields}
\alias{setTableExtraFields}
\title{Global extra table fields}
\usage{
getTableExtraFields()

setTableExtraFields(value)
}
\arguments{
\item{value}{A character vector containing the names of extra fields to include.}
}
\value{
\code{getTableExtraFields} returns the current global extra table fields.

\code{setTableExtraFields} will set the current global extra table fields and return \code{NULL} invisibly.
}
\description{
Get or set the names of the extra fields to include in a table.
}
\details{
These utilities allow users to easily set the extra table fields for all \linkS4class{DynamicMarkerTable}s at once.
Note that they work during construction of each class instance and has no effect on the \code{iSEE} application once it starts.
}
\examples{
old <- getTableExtraFields()

setTableExtraFields(LETTERS)
getTableExtraFields()

setTableExtraFields(old)
}
\author{
Aaron Lun
}
