% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot.R
\docType{methods}
\name{MAPlot-class}
\alias{MAPlot-class}
\alias{MAPlot}
\alias{initialize,MAPlot-method}
\alias{.refineParameters,MAPlot-method}
\alias{.defineDataInterface,MAPlot-method}
\alias{.fullName,MAPlot-method}
\alias{.panelColor,MAPlot-method}
\alias{.generateDotPlotData,MAPlot-method}
\alias{.allowableXAxisChoices,MAPlot-method}
\alias{.allowableYAxisChoices,MAPlot-method}
\alias{.prioritizeDotPlotData,MAPlot-method}
\alias{.colorByNoneDotPlotField,MAPlot-method}
\alias{.colorByNoneDotPlotScale,MAPlot-method}
\alias{.generateDotPlot,MAPlot-method}
\alias{.cacheCommonInfo,MAPlot-method}
\alias{.createObservers,MAPlot-method}
\alias{.hideInterface,MAPlot-method}
\title{The MAPlot class}
\description{
The MAPlot is a \linkS4class{RowDataPlot} subclass that is dedicated to creating a MA plot.
It retrieves the log-fold change and average abundance and creates a row-based plot where each point represents a feature.
Users are expected to load relevant statistics into the \code{\link{rowData}} of a \linkS4class{SummarizedExperiment}.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualization:
\itemize{
\item \code{PValueField}, a string specifying the field of \code{\link{rowData}} containing the p-values.
See \code{?\link{.acceptablePValueFields}} for more details.
\item \code{PValueThreshold}, a numeric scalar in (0, 1] specifying the threshold to use on the (adjusted) p-value.
Defaults to 0.05.
\item \code{LogFCThreshold}, a non-negative numeric scalar specifying the threshold to use on the log-fold change.
Defaults to 0.
\item \code{PValueCorrection}, a string specifying the multiple testing correction to apply.
Defaults to \code{"BH"}, but can take any value from \code{\link{p.adjust.methods}}.
}

In addition, this class inherits all slots from its parent \linkS4class{RowDataPlot},
\linkS4class{RowDotPlot}, \linkS4class{DotPlot} and \linkS4class{Panel} classes.
}

\section{Constructor}{

\code{MAPlot(...)} creates an instance of a MAPlot class,
where any slot and its value can be passed to \code{...} as a named argument.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{RowDataPlot} class.
Refer to the documentation for each method for more details on the remaining arguments.

For setting up data values:
\itemize{
\item \code{\link{.cacheCommonInfo}(x)} adds a \code{"MAPlot"} entry containing \code{pval.rowData.names}, \code{ave.rowData.names} and \code{lfc.rowData.names}.
Each of these is a character vector of permissible names for p-values, average abundances and log-fold changes, respectively;
see \code{?\link{.acceptablePValueFields}} for details.
This will also call the equivalent \linkS4class{RowDataPlot} method.
\item \code{\link{.refineParameters}(x, se)} returns \code{x} after setting \code{XAxis="Row data"}.
This will also call the equivalent \linkS4class{RowDataPlot} method for further refinements to \code{x}.
If valid p-value and log-fold change fields are not available, \code{NULL} is returned instead.
}

For defining the interface:
\itemize{
\item \code{\link{.defineDataInterface}(x, se, select_info)} returns a list of interface elements for manipulating all slots described above.
\item \code{\link{.panelColor}(x)} will return the specified default color for this panel class.
\item \code{\link{.allowableXAxisChoices}(x, se)} returns a character vector specifying the acceptable average abundance-related variables in \code{\link{rowData}(se)} that can be used as choices for the x-axis, see \code{?\link{.acceptableAveAbFields}}.
\item \code{\link{.allowableYAxisChoices}(x, se)} returns a character vector specifying the acceptable log-fold change-related variables in \code{\link{rowData}(se)} that can be used as choices for the y-axis, see \code{?\link{.acceptableLogFCFields}}.
\item \code{\link{.hideInterface}(x, field)} will return \code{TRUE} for \code{field="XAxis"},
otherwise it will call the \linkS4class{RowDataPlot} method.
\item \code{\link{.fullName}(x)} will return \code{"MA plot"}.
}

For monitoring reactive expressions:
\itemize{
\item \code{\link{.createObservers}(x, se, input, session, pObjects, rObjects)} sets up observers for all new slots described above, as well as in the parent classes via the \linkS4class{RowDataPlot} method.
}

For creating the plot:
\itemize{
\item \code{\link{.generateDotPlotData}(x, envir)} will create a data.frame of row metadata variables in \code{envir}.
This should contain average abundances on the x-axis and log-fold changes on the y-axis,
in addition to an extra field specifying whether or not the feature was considered to be significantly up or down.
The method will return the commands required to do so as well as a list of labels.
\item \code{\link{.prioritizeDotPlotData}(x, envir)} will create variables in \code{envir} marking the priority of points.
Significant features receive higher priority (i.e., are plotted over their non-significant counterparts) and are less aggressively downsampled when \code{Downsample=TRUE}.
The method will return the commands required to do this as well as a logical scalar indicating that rescaling of downsampling resolution is performed.
\item \code{\link{.colorByNoneDotPlotField}(x)} will return a string specifying the field of the data.frame (generated by \code{\link{.generateDotPlotData}}) containing the significance information.
This is to be used for coloring when \code{ColorBy="None"}.
\item \code{\link{.colorByNoneDotPlotScale}(x)} will return a string containing a \pkg{ggplot2} command to add a default color scale when \code{ColorBy="None"}.
\item \code{\link{.generateDotPlot}(x, labels, envir)} returns a list containing \code{plot} and \code{commands}, using the inital \linkS4class{ColumnDataPlot} \link{ggplot} and adding horizontal lines demarcating the log-fold change threshold.
}
}

\examples{
# Making up some results:
se <- SummarizedExperiment(matrix(rnorm(10000), 1000, 10))
rownames(se) <- paste0("GENE_", seq_len(nrow(se)))
rowData(se)$PValue <- runif(nrow(se))
rowData(se)$LogFC <- rnorm(nrow(se))
rowData(se)$AveExpr <- rnorm(nrow(se))

if (interactive()) {
    iSEE(se, initial=list(MAPlot()))
}

}
\seealso{
\link{RowDataPlot}, for the base class.
}
\author{
Aaron Lun
}
